+++ app/app.component.ts
@@ -1,8 +1,7 @@
 import {Component} from "@angular/core";
 import {LoginComponent} from "./login/login.component";
 import {HomepageComponent} from "./homepage/homepage.component";
-import {HeaderComponent} from "./header/header.component";
-import {Page1Component} from "./page1/page1.component";
+import {HeaderComponent} from "./common/header/header.component";
 import {ROUTER_DIRECTIVES} from "@angular/router";
 import {LoginService} from "./login/login.service";
 
@@ -10,7 +9,7 @@ import {LoginService} from "./login/login.service";
     selector: 'my-app',
     templateUrl: 'app/app.component.html',
     directives: [HeaderComponent, ROUTER_DIRECTIVES],
-    precompile: [LoginComponent, HomepageComponent, Page1Component],
+    precompile: [LoginComponent, HomepageComponent],
     providers: [LoginService]
 })
