+++ app/app.routes.ts
@@ -1,12 +1,10 @@
-import { provideRouter, RouterConfig } from '@angular/router';
-import { LoginComponent } from './login/login.component';
-
+import {provideRouter, RouterConfig} from "@angular/router";
+import {LoginComponent} from "./login/login.component";
 import {HomepageComponent} from "./homepage/homepage.component";
-import { Page1Component } from './page1/page1.component';
 import {HomepageRoutes} from "./homepage/homepage.routes";
 
 
-const routes: RouterConfig = [
+const routes:RouterConfig = [
     {
         path: '',
         redirectTo: '/login',
@@ -20,10 +18,6 @@ const routes: RouterConfig = [
         path: 'logo',
         redirectTo: '/login',
         pathMatch: 'full'
-    },
-    {
-        path: 'page1',
-        component: Page1Component
     },
     {
         path: 'homepage/:status',
