+++ app/common/constants/constants.data.ts
@@ -4,5 +4,5 @@ export class Constants {
         "commodi ducimus earum eum eveniet ex excepturi expedita facilis, itaque, modi nam porro quaerat quisquam " +
         "quo reiciendis tenetur unde, ut.";
 
-    
+
 }
