+++ app/common/header/header.component.ts
@@ -1,0 +1,12 @@
+import {Component} from "@angular/core";
+import {ROUTER_DIRECTIVES} from "@angular/router";
+
+@Component({
+    selector: 'header-component',
+    templateUrl: '../app/common/header/header-logo.html',
+    directives: [ROUTER_DIRECTIVES]
+})
+
+export class HeaderComponent {
+
+}
