+++ app/index.ejs
@@ -1,0 +1,49 @@
+<html>
+<head>
+    <base href="/">
+
+    <title>Angular 2 QuickStart</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <link href="../assets/fonts/roboto/Roboto-Regular.woff2" rel="stylesheet">
+    <link href="../assets/fonts/roboto/Roboto-Regular.woff" rel="stylesheet">
+    <link href="../assets/fonts/roboto/Roboto-Bold.eot" rel="stylesheet">
+    <link href="../assets/fonts/roboto/Roboto-Bold.ttf" rel="stylesheet">
+    <link href="../assets/fonts/roboto/Roboto-Bold.woff" rel="stylesheet">
+    <link href="../assets/fonts/roboto/Roboto-Regular.woff2" rel="stylesheet">
+
+    <!--Import materialize.css-->
+    <link type="text/css" rel="stylesheet" href="../assets/libs/materialize.min.css" media="screen,projection"/>
+    <link href="../assets/stylesheets/style.css" type="text/css" rel="stylesheet" media="screen,projection"/>
+
+    <!--Let browser know website is optimized for mobile-->
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+</head>
+<!-- 3. Display the application -->
+<body>
+
+<my-app></my-app>
+
+<!-- 1. Load libraries -->
+<!-- Polyfill(s) for older browsers -->
+<!--<script src="scripts/requirejs/require.js"></script>-->
+<script src="scripts/systemjs/dist/system.src.js"></script>
+<script type="text/javascript" src="scripts/core-js/client/shim.min.js"></script>
+<script src="scripts/zone.js/dist/zone.js"></script>
+<script src="scripts/reflect-metadata/Reflect.js"></script>
+
+<!--2. Configure SystemJS-->
+<script type="text/javascript" src="systemjs.config.js"></script>
+<script>
+    System.import('app').catch(function (err) {
+        console.error(err);
+    });
+</script>
+
+<!--Import jQuery before materialize.js-->
+<script type="text/javascript" src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
+<script type="text/javascript" src="../assets/libs/materialize.min.js"></script>
+
+</body>
+</html>
