+++ app/login/login.data.ts
@@ -1,0 +1,5 @@
+export class LoginData {
+    constructor(public username:string,
+                public password:string) {
+    }
+}
