+++ app/login/login.html
@@ -1,8 +1,7 @@
 <div class="container valign">
     <div class="row">
         <div class="col m5 offset-m1 s12">
-            <div class="text text-title"><a [routerLink]="['/page1']"
-                                            style="color: black">{{constants.PROJECT_NAME}}</a></div>
+            <div class="text text-title">{{constants.PROJECT_NAME}}</div>
             <div class="text text-subhead">{{constants.PROJECT_DESCRIPTION}}</div>
         </div>
         <div class="col m5 offset-m1 s12">
