+++ app/main.ts
@@ -1,5 +1,5 @@
 import {bootstrap} from "@angular/platform-browser-dynamic";
-import {Constants} from "../app/constants/constants.data";
+import {Constants} from "./common/constants/constants.data";
 import {AppComponent} from "./app.component";
 import {HTTP_PROVIDERS} from "@angular/http";
 import {APP_ROUTER_PROVIDERS} from "./app.routes";
@@ -7,3 +7,4 @@ import {APP_ROUTER_PROVIDERS} from "./app.routes";
 bootstrap(AppComponent, [
     HTTP_PROVIDERS, APP_ROUTER_PROVIDERS, Constants
 ]);
+
