--- public/app/header/header.component.ts
@@ -1,16 +1,0 @@
-/**
- * Created by alexandra on 8.7.16.
- */
-
-import { Component } from '@angular/core';
-import { ROUTER_DIRECTIVES } from '@angular/router';
-
-@Component({
-    selector: 'header-component',
-    templateUrl: 'app/header/header-logo.html',
-    directives: [ROUTER_DIRECTIVES]
-})
-
-export class HeaderComponent {
-
-}
