--- public/app/login/login.data.ts
@@ -1,6 +1,0 @@
-export class LoginData {
-    constructor(
-        public username: string,
-        public password: string
-    ) {  }
-}
