--- public/app/page1/page1.component.ts
@@ -1,15 +1,0 @@
-/**
- * Created by alexandra on 8.7.16.
- */
-
-import { Component } from '@angular/core';
-import { ROUTER_DIRECTIVES } from '@angular/router';
-
-@Component({
-    templateUrl: 'app/page1/page1.html',
-    directives: [ROUTER_DIRECTIVES]
-})
-
-export class Page1Component {
-
-}
