--- public/app/page1/page1.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Page1</title>
-</head>
-<body>
-    <h1>Page 1 for router testing</h1>
-</body>
-</html>
